Improve extension fun
